Init `oct-orchestrator` crate
Move non-CLI functionality from `oct-cli` to `oct-orchestrator`
