+++ crates/oct-cli/Cargo.toml
@@ -4,20 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-oct-cloud = { workspace = true }
+oct-orchestrator = { workspace = true }
 
 assert_cmd = { workspace = true }
 clap = { workspace = true }
 tokio = { workspace = true }
 predicates = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tempfile = { workspace = true }
-log = { workspace = true }
 env_logger = { workspace = true }
-toml = { workspace = true }
-reqwest = { workspace = true }
-mockito = { workspace = true }
 
 [lints]
 workspace = true
