+++ crates/oct-orchestrator/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "oct-orchestrator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+oct-cloud = { workspace = true }
+
+log = { workspace = true }
+mockito = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tempfile = { workspace = true }
+tokio = { workspace = true }
+toml = { workspace = true }
+
+[lints]
+workspace = true
